+++ HISTORY.md
@@ -8,7 +8,7 @@
 * Added support for returning campaign name as part of sent, draft and scheduled campaign endpoints
 * Added support for returning campaign name as part of campaign summary endpoint
 * Added sample script for get client tags endpoint
-* Breaking: Following endpoints now return ListJoinedDate for each subscriber.  Also, Date has been renamed to StateChangedDate:
+* Adding support for returning ListJoinedDate for each subscriber. 
   * List.Active()
   * List.Bounced()
   * List.Unsubscribed()